[CHORE/VG-316] 캐스팅 명시적으로 변경

+++ Source/GA6thFinal_Framework/GameScripts/Scripts/RevelationSystem/RevelationSystem.cpp
@@ -98,7 +98,7 @@ void RevelationSystem::RemoveAllExtinctionElements()
         if (garde != RevelationGrade::EXTINCTION)
             break;
 
-        RemovePlayerElement(lastIndex);
+        RemovePlayerElement(static_cast<int>(lastIndex));
         size = _playerElementList.size();
     }
 }
